+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 Notable changes to this project will be documented in this file
 
+- [X] Fix searching by username
+- [X] Add ability to search by username on the users endpoint - 7/25/22
+- [X] Display all Users if user is an Admin - 7/24/22
+  - Fix how local storage token was set
+- [X] Refactor Frontend UI - 7/23/22
+  - Use SCSS Themeing
+- [X] Refactor Custom Validators - 7/22/22
+  - Abstract String Utils
+- [X] Add custom validators to Register Form - 7/21/22
+- [X] Add Login & Register Forms - 7/20/22
+- [X] Add Angular Frontend & Connect to Nest - 7/15/22
 - [X] Prevent users from setting or updating their own roles - 7/13/22
 - [X] Add pagination to the users get all endpoint - 7/13/22
 - [X] Add a update user role endpoint & add guards - 7/13/22